backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.0.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Redis
redis>=5.0.0
celery>=5.3.0

# Vector database
qdrant-client>=1.7.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# AI/ML
google-generativeai>=0.4.0
openai>=1.10.0
anthropic>=0.18.0
litellm>=1.20.0
tiktoken>=0.5.0

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.0
unstructured>=0.12.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
structlog>=24.1.0
tenacity>=8.2.0

# Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
sentry-sdk[fastapi]>=1.39.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Linting
ruff>=0.1.0
mypy>=1.8.0
